fix integer coercion

+++ modulo-dyn.cpp
@@ -7,10 +7,10 @@
 /// Parent:
 template <unsigned* m> struct Dmint {
     Dmint():val(0){}
-    Dmint(int raw):val((raw%*m+*m)%*m){}
-    Dmint(long long raw):val((raw%*m+*m)%*m){}
+    Dmint(int raw):val((raw%(int)*m+(int)*m)%(int)*m){}
+    Dmint(long long raw):val((raw%(long long)*m+(long long)*m)%(long long)*m){}
     Dmint(unsigned raw):val(raw%*m){}
-    Dmint(unsigned long long raw):val(raw%*m){}
+    Dmint(unsigned long long raw):val(raw%(unsigned long long)*m){}
     friend Dmint operator+(Dmint a, Dmint b) { return a.val + b.val; }
     friend Dmint operator-(Dmint a, Dmint b) { return m + a.val - b.val; }
     friend Dmint operator*(Dmint a, Dmint b) { return 1ull * a.val * b.val; }
